resolved build issue

+++ src/misc/MiscFunctions_.hpp
@@ -3,7 +3,7 @@
 #include <QString>
 #include <kmt/Core.hpp>
 
-inline km::KDate to_kdate(uint8_t year, uint8_t month, uint8_t day)
+inline km::KDate to_kdate(uint16_t year, uint8_t month, uint8_t day)
 {
     return km::KDate{year,month,day};
 }
